Issue a warning on missing @key definition
When using includes you might want to pre-define identities. The user
can decide if (s)he wants to add those identities to his/her sedge
configuration. A warning is trown when keys are missing in stead of
terminating sedge, the identity parsing will be skipped and a valid ssh
config is generated.